Multi Consolidation Oracle Network (#50)
* StakeStarOracle Multi Consensus implemented
StakeStarOracle Multi Consensus Strict (with balance control) implemented

* refactoring

* set stakeStarOracleStrict as default IOracleNetwork instead of stakeStarOracle

* typo

* Fix precision in test

* Enforce typing

---------

+++ contracts/interfaces/IOracleNetwork.sol
@@ -5,5 +5,5 @@ interface IOracleNetwork {
     function latestTotalBalance()
         external
         view
-        returns (uint256 totalBalance, uint256 timestamp);
+        returns (uint256 totalBalance, uint64 timestamp);
 }
